Code cleanup (mainly removing the reason for some compiler warnings).
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1540 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/base/buffer.c
@@ -448,7 +448,7 @@ int GWEN_Buffer__FillBuffer_IoLayer(GWEN_BUFFER *bf){
 			       toread,
 			       GWEN_IO_REQUEST_FLAGS_READALL,
 			       0, 30000);
-    if (rv<toread) {
+    if (rv<(int)toread) {
       if (rv<0) {
 	DBG_INFO(GWEN_LOGDOMAIN, "here (%d)", rv);
         return rv;
